- Azure_eu and include_metadata support added - Live preview code updated - Security file added
